refactor: Implement social events hook
- move logic related to socket event handling to custom hook `useSocial` to handle defining, subscribing and unsubscribing to socket events from `UserProfile.jsx` to `useSocial.js`.
- integrate `useSocial` with required parameters in `UserProfile.jsx`.
